requirements.txt
#### ENERGI Requirements ###
torch
Pillow
numpy
pyyaml
IPython
torchsummary
tensorboard
seaborn
matplotlib